+++ phonemizer/main.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
-# coding: utf-8
-
-# Copyright 2015-2018 Mathieu Bernard
+# Copyright 2015-2019 Mathieu Bernard
 #
 # This file is part of phonemizer: you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -23,12 +21,12 @@ import logging
 import pkg_resources
 import sys
 
-from phonemizer import phonemize, separator
+from phonemizer import phonemize, separator, version
 from phonemizer.backend import (
     EspeakBackend, FestivalBackend, SegmentsBackend)
 
 
-class CatchExceptions(object):
+class CatchExceptions(object):  # pragma: nocover
     """Decorator wrapping a function in a try/except block
 
     When an exception occurs, display a user friendly message on
@@ -151,77 +149,92 @@ Exemples:
     # general arguments
     parser.add_argument(
         '--version', action='store_true',
-        help='show version information and exit')
+        help='show version information and exit.')
 
-    parser.add_argument(
+    group = parser.add_mutually_exclusive_group()
+    group.add_argument(
         '-v', '--verbose', action='store_true',
-        help='write some log messages to stderr')
+        help='write all log messages to stderr '
+        '(displays only warnings by default).')
+    group.add_argument(
+        '-q', '--quiet', action='store_true',
+        help='do not display any log message, even warnings.')
 
     parser.add_argument(
         '-j', '--njobs', type=int, metavar='<int>', default=1,
-        help='number of parallel jobs, default is %(default)s')
+        help='number of parallel jobs, default is %(default)s.')
 
     # input/output arguments
     group = parser.add_argument_group('input/output')
     group.add_argument(
         'input', default=sys.stdin, nargs='?', metavar='<file>',
-        help='input text file to phonemize, if not specified read from stdin')
+        help='input text file to phonemize, if not specified read from stdin.')
 
     group.add_argument(
         '-o', '--output', default=sys.stdout, metavar='<file>',
-        help='output text file to write, if not specified write to stdout')
+        help='output text file to write, if not specified write to stdout.')
 
     group = parser.add_argument_group('separators')
     group.add_argument(
         '-p', '--phone-separator', metavar='<str>',
         default=separator.default_separator.phone,
-        help='phone separator, default is "%(default)s"')
+        help='phone separator, default is "%(default)s".')
 
     group.add_argument(
         '-w', '--word-separator', metavar='<str>',
         default=separator.default_separator.word,
-        help='word separator, default is "%(default)s"')
+        help='word separator, default is "%(default)s".')
 
     group.add_argument(
         '-s', '--syllable-separator', metavar='<str>',
         default=separator.default_separator.syllable,
-        help='''syllable separator is available only for the festival backend,
+        help='''syllable separator, only valid for festival backend,
         this option has no effect if espeak or segments is used.
-        Default is "%(default)s"''')
+        Default is "%(default)s".''')
 
     group.add_argument(
         '--strip', action='store_true',
-        help='removes the end separators in phonemized tokens')
+        help='removes the end separators in phonemized tokens.')
 
-    group = parser.add_argument_group('language')
-
+    group = parser.add_argument_group('backends')
     group.add_argument(
         '-b', '--backend', metavar='<str>', default='espeak',
         choices=['espeak', 'festival', 'segments'],
         help="""the phonemization backend, must be 'espeak', 'festival' or
-        'segments'. Default is %(default)s""")
+        'segments'. Default is %(default)s.""")
 
+    group = parser.add_argument_group('specific to espeak backend')
+    group.add_argument(
+        '--with-stress', action='store_true',
+        help='''when the option is set, the stresses on phonemes are present
+        (stresses characters are ˈ'ˌ). By default stresses are removed.''')
+    group.add_argument(
+        '--sampa', action='store_true',
+        help='''only valid for espeak-ng and NOT supported for espeak, use the
+        "sampa" (Speech Assessment Methods Phonetic Alphabet) alphabet instead
+        of "ipa" (International Phonetic Alphabet).''')
+    group.add_argument(
+        '--language-switch', default='remove-flags',
+        choices=['keep-flags', 'remove-flags', 'remove-utterance'],
+        help="""espeak can pronounce some words in another language (typically
+        English) when phonemizing a text. This option setups the policy to use
+        when such a language switch occurs. Three values are available:
+        'keep-flags', 'remove-flags' (the default) or 'remove-utterance'. The
+        'keep-flags' policy keeps the language switching flags, for example
+        (en) or (jp), in the output. The 'remove-flags' policy removes them and
+        the 'remove-utterance' policy removes the whole line of text including
+        a language switch.""")
+
+    group = parser.add_argument_group('language')
     group.add_argument(
         '-l', '--language', metavar='<str|file>', default='en-us',
         help='''the language code of the input text, see below for a list of
         supported languages. According to the language code you
         specify, the appropriate backend (segments, espeak or
         festival) will be called in background. Default is
-        %(default)s''')
+        %(default)s.''')
 
     return parser.parse_args()
-
-
-def version():
-    """Return version information for front and backends"""
-    version = ('phonemizer-'
-               + pkg_resources.get_distribution('phonemizer').version)
-
-    return version + '\navailable backends: ' + ', '.join(
-        ('festival-' + FestivalBackend.version(),
-         ('espeak-' + ('ng-' if EspeakBackend.is_espeak_ng() else '')
-          + EspeakBackend.version()),
-         'segments-' + SegmentsBackend.version()))
 
 
 @CatchExceptions
@@ -230,20 +243,21 @@ def main():
     args = parse_args()
 
     if args.version:
-        print(version())
+        print(version.version())
         return
 
-    # configure logging according to --verbose option. If verbose,
-    # init a logger to output on stderr. Else init a logger going to
-    # the void.
+    # configure logging according to --verbose option. Init a logger to output
+    # on stderr. By default log only warnings, if verbose log all messages, if
+    # quiet do not log anything.
     logger = logging.getLogger()
     logger.handlers = []
-    logger.setLevel(logging.DEBUG)
+    handler = logging.StreamHandler(sys.stderr)
+    logger.setLevel(logging.WARNING)
     if args.verbose:
-        handler = logging.StreamHandler(sys.stderr)
-        handler.setFormatter(logging.Formatter('%(message)s'))
-    else:
+        logger.setLevel(logging.DEBUG)
+    if args.quiet:
         handler = logging.NullHandler()
+    handler.setFormatter(logging.Formatter('[%(levelname)s] %(message)s'))
     logger.addHandler(handler)
 
     # configure input as a readable stream
@@ -274,12 +288,20 @@ def main():
 
     # phonemize the input text
     out = phonemize.phonemize(
-        text, language=args.language, backend=args.backend,
-        separator=sep, strip=args.strip, njobs=args.njobs, logger=logger)
+        text,
+        language=args.language,
+        backend=args.backend,
+        separator=sep,
+        strip=args.strip,
+        with_stress=args.with_stress,
+        use_sampa=args.sampa,
+        language_switch=args.language_switch,
+        njobs=args.njobs,
+        logger=logger)
 
     if len(out):
         streamout.write(out + '\n')
 
 
-if __name__ == '__main__':
-        main()
+if __name__ == '__main__':  # pragma: nocover
+    main()