+++ test/test_espeak.py
@@ -1,6 +1,4 @@
-# coding: utf-8
-
-# Copyright 2017-2018 Mathieu Bernard
+# Copyright 2015-2019 Mathieu Bernard
 #
 # This file is part of phonemizer: you can redistribute it and/or
 # modify it under the terms of the GNU General Public License as
@@ -45,6 +43,16 @@ def test_english():
     assert out == u'həloʊ wɜːld\nɡʊdbaɪ\nθɜːd laɪn\njɛt ɐnʌðɚ'
 
 
+def test_stress():
+    backend = EspeakBackend('en-us', with_stress=False)
+    assert u'həloʊ wɜːld' == backend._phonemize_aux(
+        u'hello world', separator.default_separator, True)[0]
+
+    backend = EspeakBackend('en-us', with_stress=True)
+    assert u'həlˈoʊ wˈɜːld' == backend._phonemize_aux(
+        u'hello world', separator.default_separator, True)[0]
+
+
 def test_french():
     backend = EspeakBackend('fr-fr')
     text = u'bonjour le monde'
@@ -52,3 +60,54 @@ def test_french():
     expected = [u'b ɔ̃ ʒ u ʁ ;eword l ə- ;eword m ɔ̃ d ;eword ']
     out = backend._phonemize_aux(text, sep, False)
     assert out == expected
+
+
+@pytest.mark.skipif(
+    not EspeakBackend.is_espeak_ng(),
+    reason='Arabic is only supported by espeak-ng')
+def test_arabic():
+    backend = EspeakBackend('ar')
+    text = u'السلام عليكم'
+    sep = separator.Separator()
+
+    # Arabic seems to have changed starting at espeak-ng-1.49.3
+    if tuple(EspeakBackend.version().split('.')) >= ('1', '49', '3'):
+        expected = [u'ʔassalaːm ʕliːkm ']
+    else:
+        expected = [u'ʔassalaam ʕaliijkum ']
+    out = backend._phonemize_aux(text, sep, False)
+    assert out == expected
+
+
+def test_language_switch():
+    text = '\n'.join([
+        "j'aime l'anglais",
+        "j'aime le football",
+        "football",
+        "surtout le real madrid",
+        "n'utilise pas google"])
+
+    backend = EspeakBackend('fr-fr', language_switch='keep-flags')
+    out = backend._phonemize_aux(text, separator.Separator(), True)
+    assert out == [
+        'ʒɛm lɑ̃ɡlɛ',
+        'ʒɛm lə- (en)fʊtbɔːl(fr)',
+        '(en)fʊtbɔːl(fr)',
+        'syʁtu lə- (en)ɹiəl(fr) madʁid',
+        'nytiliz pa (en)ɡuːɡəl(fr)']
+
+    backend = EspeakBackend('fr-fr', language_switch='remove-flags')
+    out = backend._phonemize_aux(text, separator.Separator(), True)
+    assert out == [
+        'ʒɛm lɑ̃ɡlɛ',
+        'ʒɛm lə- fʊtbɔːl',
+        'fʊtbɔːl',
+        'syʁtu lə- ɹiəl madʁid',
+        'nytiliz pa ɡuːɡəl']
+
+    backend = EspeakBackend('fr-fr', language_switch='remove-utterance')
+    out = backend._phonemize_aux(text, separator.Separator(), True)
+    assert out == ['ʒɛm lɑ̃ɡlɛ']
+
+    with pytest.raises(RuntimeError):
+        backend = EspeakBackend('fr-fr', language_switch='foo')